refactor(*): example structure update
