+++ example/lib/data/data_sources/restorable_dessert_selections.dart
@@ -1,0 +1,43 @@
+import 'package:example/data/models/dessert.dart';
+import 'package:flutter/material.dart';
+
+class RestorableDessertSelections extends RestorableProperty<Set<int>> {
+  Set<int> _dessertSelections = {};
+
+  /// Returns whether or not a dessert row is selected by index.
+  bool isSelected(int index) => _dessertSelections.contains(index);
+
+  /// Takes a list of [Dessert]s and saves the row indices of selected rows
+  /// into a [Set].
+  void setDessertSelections(List<Dessert> desserts) {
+    final updatedSet = <int>{};
+    for (var i = 0; i < desserts.length; i += 1) {
+      var dessert = desserts[i];
+      if (dessert.selected) {
+        updatedSet.add(i);
+      }
+    }
+    _dessertSelections = updatedSet;
+    notifyListeners();
+  }
+
+  @override
+  Set<int> createDefaultValue() => _dessertSelections;
+
+  @override
+  Set<int> fromPrimitives(Object? data) {
+    final selectedItemIndices = data as List<dynamic>;
+    _dessertSelections = {
+      ...selectedItemIndices.map<int>((dynamic id) => id as int),
+    };
+    return _dessertSelections;
+  }
+
+  @override
+  void initWithValue(Set<int> value) {
+    _dessertSelections = value;
+  }
+
+  @override
+  Object toPrimitives() => _dessertSelections.toList();
+}
