+++ example/lib/data_table.dart
@@ -1,7 +1,8 @@
+import 'package:example/data/data_sources/dessert_data_source.dart';
+import 'package:example/data/data_sources/restorable_dessert_selections.dart';
+import 'package:example/data/models/dessert.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/rendering.dart';
-
-import 'data_source.dart';
 
 // Copyright 2019 The Flutter team. All rights reserved.
 // Use of this source code is governed by a BSD-style license that can be
