+++ example/lib/data_table2_scrollup.dart
@@ -1,8 +1,8 @@
 import 'package:data_table_2/data_table_2.dart';
+import 'package:example/data/data_sources/dessert_data_source.dart';
+import 'package:example/data/models/dessert.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/rendering.dart';
-
-import 'data_source.dart';
 
 // Copyright 2019 The Flutter team. All rights reserved.
 // Use of this source code is governed by a BSD-style license that can be
