+++ lib/src/paginated_data_tables/paginated_data_table_2.dart
@@ -1,5 +1,3 @@
-library data_table_2;
-
 // Copyright 2014 The Flutter Authors. All rights reserved.
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
